order sales and test refactoring

+++ src/common/dto/entities/production/product-inventory.dto.ts
@@ -1,11 +1,11 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, Int, ObjectType } from '@nestjs/graphql';
 
 @ObjectType('ProductInventory')
 export class ProductInventory {
-    @Field(() => Int, { nullable: true })
+    @Field(() => Float, { nullable: true })
     kilos: number | null;
 
-    @Field(() => Int, { nullable: true })
+    @Field(() => Float, { nullable: true })
     groups: number | null;
 
     @Field(() => String, { nullable: true })
